+++ eio-okta-sync/src/okta/snapshot.rs
@@ -1,4 +1,5 @@
 use std::collections::BTreeMap;
+use std::path::Path;
 
 use eio_okta_data::v2024_07_0::management::components::schemas::{Group, User};
 
@@ -7,4 +8,12 @@ pub struct Snapshot {
   pub groups: Vec<Group>,
   pub group_users: BTreeMap<String, Vec<User>>,
   pub users: Vec<User>,
+}
+
+impl Snapshot {
+  pub(crate) fn read_from_file(path: impl AsRef<Path>) -> Result<Self, crate::Error> {
+    let yaml = fs_err::read_to_string(path)?;
+    let this = serde_yml::from_str(&yaml)?;
+    Ok(this)
+  }
 }